Alipour/Roche Version
This is the version of the code that was developed by Ryan Alipour and Mr. Roche, based on Kevin Wu and Julie Geng's original code, and was used to operate the robot in Day 2 of the BAE Granite State Regional. It was current as of the end of the day on March 2, 2013.

It includes an autonomous command that will (we think) accurately and reliably fire 2-3 frisbees from any corner of the pyramid (with suitable modifications).

Signed-off-by: Seth Battis <SethBattis@stmarksschool.org>